Move routes to their own component
